move tests, check if dynamic clients are available in getclients
